% Generated by roxygen2 (4.1.0.9001): do not edit by hand
\name{builtins}
\alias{builtins}
\alias{colMaxsBy}
\alias{colMeansBy}
\alias{colMediansBy}
\alias{colMinsBy}
\alias{colSumsBy}
\alias{rowMaxsBy}
\alias{rowMeansBy}
\alias{rowMediansBy}
\alias{rowMinsBy}
\alias{rowSumsBy}
\title{Built-in Group Apply Utility Functions}
\usage{
colSumsBy(x, BY, ...)

rowSumsBy(x, BY, ...)

rowMeansBy(x, BY, ...)

colMeansBy(x, BY, ...)

rowMediansBy(x, BY, ...)

colMediansBy(x, BY, ...)

rowMaxsBy(x, BY, ...)

colMaxsBy(x, BY, ...)

rowMinsBy(x, BY, ...)

colMinsBy(x, BY, ...)
}
\arguments{
  \item{x}{matrix-like object on which \code{\link{apply}} can be called.}

  \item{BY}{factor or object coerced to a factor, that defines the groups within
which the function \code{FUN} is applied.

If \code{x} is an ExpressionSet object, then \code{BY} can be the names of a
sample (resp. feature) annotation variable if \code{MARGIN=1} (resp. \code{MARGIN=2L})
(see examples).}

  \item{...}{extra parameters passed to \code{FUN}.}
}
\description{
  These functions provide convenience shortcut for commonly
performed group-apply operations, such as column/row group means, medians,
maxs, mins, etc...

\code{col<STAT>By} computes for each column a given statistic within separate groups of rows, which are defined by a factor.

\code{row<STAT>By} computes for each row a given statistic within separate groups of columns, which are defined by a factor.
}

